+++ src/main/java/kitchenpos/order/common/application/OrderService.java
@@ -3,6 +3,9 @@ package kitchenpos.order.common.application;
 import kitchenpos.menu.domain.model.Menu;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.order.common.domain.*;
+import kitchenpos.order.common.domain.model.Order;
+import kitchenpos.order.common.domain.model.OrderLineItem;
+import kitchenpos.order.common.domain.repository.OrderRepository;
 import kitchenpos.order.deliveryorder.infra.KitchenridersClient;
 import kitchenpos.order.eatinorder.domain.OrderTable;
 import kitchenpos.order.eatinorder.domain.OrderTableRepository;
