+++ src/main/java/kitchenpos/order/common/domain/model/Order.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.common.domain;
+package kitchenpos.order.common.domain.model;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -12,6 +12,8 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
+import kitchenpos.order.common.domain.OrderStatus;
+import kitchenpos.order.common.domain.OrderType;
 import kitchenpos.order.eatinorder.domain.OrderTable;
 
 import java.time.LocalDateTime;
