+++ src/main/java/kitchenpos/order/common/infra/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order.common.domain;
+package kitchenpos.order.common.infra;
 
+import kitchenpos.order.common.domain.model.Order;
+import kitchenpos.order.common.domain.repository.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
