+++ src/main/java/kitchenpos/order/eatinorder/application/OrderTableService.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.eatinorder.application;
 
-import kitchenpos.order.common.domain.OrderRepository;
+import kitchenpos.order.common.domain.repository.OrderRepository;
 import kitchenpos.order.common.domain.OrderStatus;
 import kitchenpos.order.eatinorder.domain.OrderTable;
 import kitchenpos.order.eatinorder.domain.OrderTableRepository;
