+++ src/main/java/kitchenpos/product/infra/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.infra;
 
+import kitchenpos.product.domain.model.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
