+++ src/test/java/kitchenpos/order/common/application/InMemoryOrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.common.application;
 
-import kitchenpos.order.common.domain.Order;
-import kitchenpos.order.common.domain.OrderRepository;
+import kitchenpos.order.common.domain.model.Order;
+import kitchenpos.order.common.domain.repository.OrderRepository;
 import kitchenpos.order.common.domain.OrderStatus;
 import kitchenpos.order.eatinorder.domain.OrderTable;
